requirements.txt
appdirs==1.4.3
Cython==0.25.2
numpy==1.12.1
olefile==0.44
opencv-python==3.2.0.7
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
pytesseract==0.1.6
six==1.10.0
Wand==0.4.4
nltk==3.2.2
probablepeople==0.5.2
